requirements.txt
# Core Framework
streamlit>=1.28.0

# Environment Variables
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# AI & ML
google-generativeai>=0.3.0
openai>=1.0.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pytesseract>=0.3.0

# Document Processing
python-docx>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization
plotly>=5.0.0

# Web Scraping & Requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Authentication & Security
bcrypt>=4.0.0

# Utilities
schedule>=1.2.0

# Streamlit Components
streamlit-lottie>=0.0.5
streamlit-option-menu>=0.3.0
streamlit-extras>=0.3.0





# Mock Interview
# Note: VAPI runs via browser JS (CDN), no Python package needed
